// src/auth/enums/role.enum.ts
export enum Role {
  CURRICULUM_SELLER = 'CURRICULUM_SELLER',
  CURRICULUM_ADMIN = 'CURRICULUM_ADMIN',
  COURSE_CREATOR_ADMIN = 'COURSE_CREATOR_ADMIN',
  INSTRUCTOR_ADMIN = 'INSTRUCTOR_ADMIN',
  BOOKING_PROFESSIONAL = 'BOOKING_PROFESSIONAL',
  BOOKING_USER = 'BOOKING_USER',
  STUDENT = 'STUDENT',
  GUEST_USER = 'GUEST_USER',
  ADMIN = 'ADMIN',
}